+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [0.0.2]
 
+* `enableEdgeDragGesture` and `edgeDragWidth` was added. Thanks Brayan Cantos.
 * `position` property and `SideMenuPosition` were added, now we can display the menu from the left or right. Thanks Jorge Alejandro.
-
 
 ## [0.0.1 - 0.0.1+2]